requirements.txt
requests
pandas==1.5.2
numpy==1.22.4
matplotlib
jupyter
scikit-learn
python-binance
websocket-client
trendln
plotly
mplfinance
statsmodels
cvxpy
pyfolio
backtesting
pandas_ta
fracdiff
ta
xgboost
seaborn
doubtlab
lazy_object_proxy
pyts
lightgbm
nbconvert
ngboost
imblearn
ccxt
